finished making c++ types explicit

+++ include/SegmentedSieve.hpp
@@ -4,14 +4,15 @@ using namespace std;
 
 class SegmentedSieve : public SieveTemplate<bool> {
 private:
-    long x, y, z;
+    uint32_t x, y, z;
+    // TODO: include w variable to get honest rectangle
 
 public:
-    SegmentedSieve(long, long, long, bool = true);
+    SegmentedSieve(uint32_t, uint32_t, uint32_t, bool = true);
     // overriding virtual methods
     void setSmallPrimes() override;
     void setSieveArray() override;
     void crossOffMultiples(gint) override;
     void setBigPrimes() override;
-    unsigned long getCountBigPrimes() override;
+    uint64_t getCountBigPrimes() override;
 };
